# docs/system_manager_openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/cache/download": {
      "post": {
        "tags": [
          "cache"
        ],
        "summary": "Start download for a chute",
        "operationId": "download_cache_download_post",
        "parameters": [
          {
            "name": "force",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Re-download if already present, and discard a partial download owned by another process (e.g. a chute pod) before re-downloading",
              "default": false,
              "title": "Force"
            },
            "description": "Re-download if already present, and discard a partial download owned by another process (e.g. a chute pod) before re-downloading"
          },
          {
            "name": "X-Chutes-Hotkey",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Chutes-Hotkey"
            }
          },
          {
            "name": "X-Chutes-Nonce",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Chutes-Nonce"
            }
          },
          {
            "name": "X-Chutes-Signature",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Chutes-Signature"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DownloadRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CacheDownloadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/cache/download/status": {
      "get": {
        "tags": [
          "cache"
        ],
        "summary": "Get download status by chute_id or all",
        "operationId": "download_status_cache_download_status_get",
        "parameters": [
          {
            "name": "chute_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional chute_id to filter",
              "title": "Chute Id"
            },
            "description": "Optional chute_id to filter"
          },
          {
            "name": "X-Chutes-Hotkey",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Chutes-Hotkey"
            }
          },
          {
            "name": "X-Chutes-Nonce",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Chutes-Nonce"
            }
          },
          {
            "name": "X-Chutes-Signature",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Chutes-Signature"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CacheDownloadStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/cache/{chute_id}": {
      "delete": {
        "tags": [
          "cache"
        ],
        "summary": "Remove cache for a chute",
        "operationId": "delete_chute_cache__chute_id__delete",
        "parameters": [
          {
            "name": "chute_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Chute Id"
            }
          },
          {
            "name": "force",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Force delete even if download is in progress",
              "default": false,
              "title": "Force"
            },
            "description": "Force delete even if download is in progress"
          },
          {
            "name": "X-Chutes-Hotkey",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Chutes-Hotkey"
            }
          },
          {
            "name": "X-Chutes-Nonce",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Chutes-Nonce"
            }
          },
          {
            "name": "X-Chutes-Signature",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Chutes-Signature"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Delete Chute Cache  Chute Id  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/cache/{chute_id}/cancel": {
      "post": {
        "tags": [
          "cache"
        ],
        "summary": "Cancel an in-progress download",
        "operationId": "cancel_download_cache__chute_id__cancel_post",
        "parameters": [
          {
            "name": "chute_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Chute Id"
            }
          },
          {
            "name": "cleanup",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Delete partial files from disk after cancelling",
              "default": false,
              "title": "Cleanup"
            },
            "description": "Delete partial files from disk after cancelling"
          },
          {
            "name": "X-Chutes-Hotkey",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Chutes-Hotkey"
            }
          },
          {
            "name": "X-Chutes-Nonce",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Chutes-Nonce"
            }
          },
          {
            "name": "X-Chutes-Signature",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Chutes-Signature"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CacheCancelResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/cache/purge": {
      "post": {
        "tags": [
          "cache"
        ],
        "summary": "Purge stale HuggingFace revisions from all cached chutes",
        "operationId": "purge_stale_cache_purge_post",
        "parameters": [
          {
            "name": "X-Chutes-Hotkey",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Chutes-Hotkey"
            }
          },
          {
            "name": "X-Chutes-Nonce",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Chutes-Nonce"
            }
          },
          {
            "name": "X-Chutes-Signature",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Chutes-Signature"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CachePurgeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/cache/cleanup": {
      "post": {
        "tags": [
          "cache"
        ],
        "summary": "Cleanup cache by age and max size",
        "operationId": "cleanup_cache_cleanup_post",
        "parameters": [
          {
            "name": "max_age_days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 5,
              "title": "Max Age Days"
            }
          },
          {
            "name": "max_size_gb",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 100,
              "title": "Max Size Gb"
            }
          },
          {
            "name": "X-Chutes-Hotkey",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Chutes-Hotkey"
            }
          },
          {
            "name": "X-Chutes-Nonce",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Chutes-Nonce"
            }
          },
          {
            "name": "X-Chutes-Signature",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Chutes-Signature"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/CleanupRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CacheCleanupResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/cache/overview": {
      "get": {
        "tags": [
          "cache"
        ],
        "summary": "List cache contents and sizes",
        "operationId": "overview_cache_overview_get",
        "parameters": [
          {
            "name": "X-Chutes-Hotkey",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Chutes-Hotkey"
            }
          },
          {
            "name": "X-Chutes-Nonce",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Chutes-Nonce"
            }
          },
          {
            "name": "X-Chutes-Signature",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Chutes-Signature"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CacheOverviewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CacheCancelResponse": {
        "properties": {
          "chute_id": {
            "type": "string",
            "title": "Chute Id",
            "description": "Chute ID"
          },
          "status": {
            "$ref": "#/components/schemas/CacheDownloadStatus",
            "description": "One of: cancelled, not_in_progress, not_found"
          }
        },
        "type": "object",
        "required": [
          "chute_id",
          "status"
        ],
        "title": "CacheCancelResponse"
      },
      "CacheChuteStatus": {
        "properties": {
          "chute_id": {
            "type": "string",
            "title": "Chute Id",
            "description": "Chute ID"
          },
          "status": {
            "$ref": "#/components/schemas/CacheChuteStatusEnum",
            "description": "One of: in_progress, present, missing, failed, incomplete, stale"
          },
          "percent_complete": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Percent Complete",
            "description": "Download progress 0-100 when in_progress and total size is known; omitted otherwise"
          },
          "download_rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Download Rate",
            "description": "Average download speed in bytes/sec for the current session; omitted when not in_progress"
          },
          "eta_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Eta Seconds",
            "description": "Estimated seconds until download completes; omitted when rate or total size is unknown"
          },
          "repo_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Repo Id",
            "description": "HF repo ID when present or in_progress"
          },
          "revision": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revision",
            "description": "Revision when present or in_progress"
          },
          "size_bytes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Size Bytes",
            "description": "Size in bytes when present"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error",
            "description": "Error message when status is failed"
          }
        },
        "type": "object",
        "required": [
          "chute_id",
          "status"
        ],
        "title": "CacheChuteStatus"
      },
      "CacheChuteStatusEnum": {
        "type": "string",
        "enum": [
          "in_progress",
          "present",
          "missing",
          "failed",
          "incomplete",
          "stale"
        ],
        "title": "CacheChuteStatusEnum",
        "description": "Status for a chute in the download status (GET) and overview."
      },
      "CacheCleanupResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status",
            "description": "Cleanup status",
            "examples": [
              "completed"
            ]
          },
          "freed_bytes": {
            "type": "integer",
            "title": "Freed Bytes",
            "description": "Bytes freed by removing entire chutes",
            "default": 0
          },
          "purged_bytes": {
            "type": "integer",
            "title": "Purged Bytes",
            "description": "Bytes freed by pruning stale HF revisions from kept chutes",
            "default": 0
          },
          "removed_chutes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Removed Chutes",
            "description": "Chute IDs removed"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "CacheCleanupResponse"
      },
      "CacheDownloadResponse": {
        "properties": {
          "chute_id": {
            "type": "string",
            "title": "Chute Id",
            "description": "Chute ID"
          },
          "status": {
            "$ref": "#/components/schemas/CacheDownloadStatus",
            "description": "One of: started, present, in_progress, failed"
          }
        },
        "type": "object",
        "required": [
          "chute_id",
          "status"
        ],
        "title": "CacheDownloadResponse"
      },
      "CacheDownloadStatus": {
        "type": "string",
        "enum": [
          "started",
          "present",
          "in_progress",
          "failed",
          "cancelled"
        ],
        "title": "CacheDownloadStatus",
        "description": "Status returned by the download (POST) and cancel (POST) endpoints."
      },
      "CacheDownloadStatusResponse": {
        "properties": {
          "chutes": {
            "items": {
              "$ref": "#/components/schemas/CacheChuteStatus"
            },
            "type": "array",
            "title": "Chutes",
            "description": "Status per chute"
          }
        },
        "type": "object",
        "required": [
          "chutes"
        ],
        "title": "CacheDownloadStatusResponse"
      },
...
      "CacheOverviewResponse": {
        "properties": {
          "total_size_bytes": {
            "type": "integer",
            "title": "Total Size Bytes",
            "description": "Total cache size in bytes"
          },
          "chutes": {
            "items": {
              "$ref": "#/components/schemas/CacheOverviewEntry"
            },
            "type": "array",
            "title": "Chutes",
            "description": "Entries per chute"
          }
        },
        "type": "object",
        "required": [
          "total_size_bytes",
          "chutes"
        ],
        "title": "CacheOverviewResponse"
      },
      "CachePurgeResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status",
            "description": "Purge status",
            "examples": [
              "completed"
            ]
          },
          "purged_bytes": {
            "type": "integer",
            "title": "Purged Bytes",
            "description": "Bytes freed by pruning stale HF revisions",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "CachePurgeResponse"
      },
      "CleanupRequest": {
        "properties": {
          "max_age_days": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Max Age Days",
            "description": "Remove entries older than this many days",
            "default": 5
          },
          "max_size_gb": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Max Size Gb",
            "description": "Target max cache size in GB",
            "default": 100
          },
          "exclude_pattern": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exclude Pattern",
            "description": "Skip repos containing this string"
          }
        },
        "type": "object",
        "title": "CleanupRequest"
      },
...
      "DownloadRequest": {
        "properties": {
          "chute_id": {
            "type": "string",
            "title": "Chute Id",
            "description": "Chute ID to download model for"
          }
        },
        "type": "object",
        "required": [
          "chute_id"
        ],
        "title": "DownloadRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}