# src/sek8s/sek8s/system_manager/cache/router.py
"""Cache submodule: FastAPI router and route handlers.

Every endpoint delegates to CacheManager / ChuteModel via
``request.app.state.cache_manager``.
"""

from __future__ import annotations

import os
import uuid
from typing import Optional

from fastapi import APIRouter, Depends, HTTPException, Query, Request

from sek8s.services.util import authorize

from .manager import CacheManager
from .models import CacheChuteStatusEnum, ChuteSnapshot, CleanupRequest, DownloadRequest
from .responses import (
    CacheCancelResponse,
    CacheChuteStatus,
    CacheCleanupResponse,
    CacheDownloadResponse,
    CacheDownloadStatus,
    CacheDownloadStatusResponse,
    CacheOverviewEntry,
    CacheOverviewResponse,
    CachePurgeResponse,
)
from .util import fetch_hf_info

router = APIRouter()


def _validate_chute_id(chute_id: str) -> str:
    """Validate chute_id is a well-formed UUID. Raises HTTPException if not."""
    try:
        uuid.UUID(chute_id)
    except (ValueError, AttributeError):
        raise HTTPException(status_code=400, detail="chute_id must be a valid UUID")
    return chute_id


def _snap_to_status(snap: ChuteSnapshot) -> CacheChuteStatus:
    return CacheChuteStatus(
        chute_id=snap.chute_id,
        status=snap.status,
        percent_complete=snap.percent_complete,
        download_rate=snap.download_rate,
        eta_seconds=snap.eta_seconds,
        repo_id=snap.repo_id or None,
        revision=snap.revision,
        size_bytes=snap.size_bytes or None,
        error=snap.error,
    )


def _snap_to_overview(snap: ChuteSnapshot) -> CacheOverviewEntry:
    return CacheOverviewEntry(
        chute_id=snap.chute_id,
        repo_id=snap.repo_id,
        revision=snap.revision,
        size_bytes=snap.size_bytes,
        last_accessed=snap.last_accessed,
        status=snap.status,
    )


async def get_cache_manager(request: Request) -> CacheManager:
    """FastAPI dependency that pulls the CacheManager off app.state."""
    return request.app.state.cache_manager


@router.post(
    "/download",
    response_model=CacheDownloadResponse,
    summary="Start download for a chute",
)
async def download(
    request: DownloadRequest,
    force: bool = Query(
        False,
        description=(
            "Re-download if already present, and discard a partial download "
            "owned by another process (e.g. a chute pod) before re-downloading"
        ),
    ),
    mgr: CacheManager = Depends(get_cache_manager),
    _auth: bool = Depends(authorize(allow_miner=True, purpose="cache")),
) -> CacheDownloadResponse:
    chute_id = _validate_chute_id(request.chute_id)

    chute = await mgr.get_or_create(chute_id)

    if chute.is_in_progress:
        return CacheDownloadResponse(
            chute_id=chute_id, status=CacheDownloadStatus.IN_PROGRESS
        )

    if chute.status == CacheChuteStatusEnum.PRESENT and not force:
        return CacheDownloadResponse(
            chute_id=chute_id, status=CacheDownloadStatus.PRESENT
        )

    try:
        info = await fetch_hf_info(chute_id)
    except HTTPException:
        raise
    repo_id = info.repo_id
    if not repo_id:
        raise HTTPException(status_code=502, detail="Validator did not return repo_id")
    revision = info.revision or "main"

    try:
        await chute.start_download(repo_id, revision, force=force)
    except ValueError as e:
        raise HTTPException(status_code=409, detail=str(e))
    return CacheDownloadResponse(chute_id=chute_id, status=CacheDownloadStatus.STARTED)


@router.get(
    "/download/status",
    response_model=CacheDownloadStatusResponse,
    summary="Get download status by chute_id or all",
)
async def download_status(
    chute_id: Optional[str] = Query(None, description="Optional chute_id to filter"),
    mgr: CacheManager = Depends(get_cache_manager),
    _auth: bool = Depends(authorize(allow_miner=True, purpose="cache")),
) -> CacheDownloadStatusResponse:
    if chute_id:
        chute = await mgr.get(chute_id)
        if chute is None:
            return CacheDownloadStatusResponse(
                chutes=[
                    CacheChuteStatus(
                        chute_id=chute_id,
                        status=CacheChuteStatusEnum.MISSING,
                        percent_complete=None,
                        download_rate=None,
                        eta_seconds=None,
                        repo_id=None,
                        revision=None,
                        size_bytes=None,
                        error=None,
                    )
                ]
            )
        return CacheDownloadStatusResponse(
            chutes=[_snap_to_status(await chute.snapshot())]
        )

    snapshots = await mgr.all_snapshots()
    return CacheDownloadStatusResponse(chutes=[_snap_to_status(s) for s in snapshots])


@router.delete(
    "/{chute_id}",
    summary="Remove cache for a chute",
)
async def delete_chute(
    chute_id: str,
    force: bool = Query(
        False, description="Force delete even if download is in progress"
    ),
    mgr: CacheManager = Depends(get_cache_manager),
    _auth: bool = Depends(authorize(allow_miner=True, purpose="cache")),
) -> dict:
    _validate_chute_id(chute_id)

    chute = await mgr.get(chute_id)
    if chute is None:
        return {"status": "ok", "message": "not found"}
    if chute.is_in_progress and not force:
        raise HTTPException(
            status_code=409, detail="Download in progress for this chute"
        )

    await mgr.remove(chute_id)
    return {"status": "ok", "message": "deleted"}


@router.post(
    "/{chute_id}/cancel",
    response_model=CacheCancelResponse,
    summary="Cancel an in-progress download",
)
async def cancel_download(
    chute_id: str,
    cleanup: bool = Query(
        False,
        description="Delete partial files from disk after cancelling",
    ),
    mgr: CacheManager = Depends(get_cache_manager),
    _auth: bool = Depends(authorize(allow_miner=True, purpose="cache")),
) -> CacheCancelResponse:
    _validate_chute_id(chute_id)

    try:
        await mgr.cancel(chute_id, cleanup=cleanup)
    except KeyError:
        raise HTTPException(status_code=404, detail="Chute not found")
    except ValueError as e:
        raise HTTPException(status_code=409, detail=str(e))
    return CacheCancelResponse(
        chute_id=chute_id,
        status=CacheDownloadStatus.CANCELLED,
    )


@router.post(
    "/purge",
    response_model=CachePurgeResponse,
    summary="Purge stale HuggingFace revisions from all cached chutes",
)
async def purge_stale(
    mgr: CacheManager = Depends(get_cache_manager),
    _auth: bool = Depends(authorize(allow_miner=True, purpose="cache")),
) -> CachePurgeResponse:
    purged_bytes = await mgr.purge_stale()
    return CachePurgeResponse(status="completed", purged_bytes=purged_bytes)


@router.post(
    "/cleanup",
    response_model=CacheCleanupResponse,
    summary="Cleanup cache by age and max size",
)
async def cleanup(
    body: Optional[CleanupRequest] = None,
    max_age_days: int = Query(5, ge=0),
    max_size_gb: int = Query(100, ge=0),
    mgr: CacheManager = Depends(get_cache_manager),
    _auth: bool = Depends(authorize(allow_miner=True, purpose="cache")),
) -> CacheCleanupResponse:
    age = body.max_age_days if body else max_age_days
    size = body.max_size_gb if body else max_size_gb
    exclude = (body.exclude_pattern if body else None) or os.environ.get(
        "CLEANUP_EXCLUDE"
    )

    result = await mgr.cleanup(age, size, exclude)
    return CacheCleanupResponse(
        status="completed",
        freed_bytes=result.freed_bytes,
        purged_bytes=result.purged_bytes,
        removed_chutes=result.removed_chutes,
    )


@router.get(
    "/overview",
    response_model=CacheOverviewResponse,
    summary="List cache contents and sizes",
)
async def overview(
    mgr: CacheManager = Depends(get_cache_manager),
    _auth: bool = Depends(authorize(allow_miner=True, purpose="cache")),
) -> CacheOverviewResponse:
    snapshots = await mgr.all_snapshots()
    entries = [_snap_to_overview(s) for s in snapshots]
    total = sum(e.size_bytes for e in entries)
    return CacheOverviewResponse(total_size_bytes=total, chutes=entries)